pyproject.toml
[project]
name = "cli-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
    "rich>=13.7.1",
]
